Add the tmux yank plugin
https://github.com/tmux-plugins/tmux-yank
Now I can select the text in tmux with `y` and it gets copied to the
clipboard and use Cmd-v(outside the terminal) or the prefix-] for paste.

I only installed this for my mac and it worked! For mac I brew installed reattach-to-user-namespace
For Linux I have yet to test this